+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### driver-native
 
+* Feature [#62](https://github.com/asfernandes/node-firebird-drivers/issues/62) - Change value of getDefaultLibraryFilename in MacOS.
 * Feature [#63](https://github.com/asfernandes/node-firebird-drivers/issues/63) - Make getMaster return OS error when failing to load the client library.
 
 ## 2.1.0 (December, 29, 2019)